+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Master (unreleased)
 
+* Add `RSpec/NamedSubject` cop. ([@backus][])
+
 ## 1.5.2 (2016-08-01)
 
 * Drop support for ruby `2.0.0` and `2.1.0`. ([@backus][])